+++ web/src/contexts/conversation/index.jsx
@@ -1,0 +1,9 @@
+import { createContext } from "react";
+
+/**
+ * messages, dispatch
+ */
+export const ConversationContext = createContext({
+    groupedConvs: {},
+    dispatch: () => { },
+});
